Added creative commons license.
At the time of this commit verified by the history and by github (force
pushes disabled) are that the contributors fell under two categories.

1. They are being employed by the University of Illinois at the Time of
Committing and billed said work, making the work owned by the university
2. Otherwise small changes.

This reliscense is to ensure that no single copyright holder can take
down large portions of the book. This license is similar to the CC
By-3.0 License which all contributors agreed to when making edits in the
original wiki.